Add robosuite experiments

+++ vla-scripts/extern/verify_openvla.py
@@ -68,7 +68,7 @@ def verify_openvla() -> None:
         MODEL_PATH,
         attn_implementation="flash_attention_2",
         torch_dtype=torch.float16,
-        quantization_config=BitsAndBytesConfig(load_in_4bit=True),
+        quantization_config=BitsAndBytesConfig(load_in_4bit=True, bnb_4bit_compute_dtype="float16"),
         low_cpu_mem_usage=True,
         trust_remote_code=True,
     )
